requirements.txt
flask
gunicorn
python-dotenv
flask_wtf
pymongo
flask-security
email_validator
passlib